Ensure that only one post retrieval interval is set at a time.
This fixes a bug where multiple intervals were set, but only the last one would ever be cleared, which resulted in new posts being retrieved (and consequently disrupting the user when the layout changed).

It happened when the user would be in the middle of a page and then refresh. init() would fire because of the refresh and call enableRetrieval(), and then the browser would automatically scroll to the top of the page, which would trigger a second call to enableRetrieval().


git-svn-id: https://plugins.svn.wordpress.org/tagregator/trunk@1010609 b8457f37-d9ea-0310-8a92-e5e31aec5664